Kesar Control Leading Manufacturer of Stability Chamber
Kesar control systems, manufacturer of pharmaceutical and scientific equipment established in the year of 2007 are here with the best people and the best manufacturing process.
To conduct any sort of experiment one of the most important things is machinery, here we provide you with high-quality and durable machines which work smoothly and along with that we give the most useful features.
Pharmaceutical and scientific experiments need a wide variety of machines to produce, store and keep experiments stable for a long period of time so it is very much important that you've to choose wisely in this department.
Stability Chamber, widely used by pharmaceutical companies, is highly in demand after the onset of the COVID-19 Pandemic. Medicines, drugs and especially vaccines need to be stored in monitored conditions. This includes monitoring of the temperature, light exposure, and moisture, amongst other conditions.
Cold chamber requires a range of cooling and heating technology to easily navigate through different temperatures based on specific requirements quickly and easily. 
For pharmaceutical and scientific requirements of storage of light-sensitive articles, Photostability Chamber is developed using state of the art technology to ensure uniform light distribution and convenient adjustment of exposure.
Walk in stability chamber is also custom-built using exceptional technology to maximize efficiency in storage by stabilizing large-scale storage requirements.
Kesar Control Systems is a trusted name when it comes to BOD incubator  and Laboratory Incubator manufacturing.The various uses include but are not limited to vaccine preservation, life cycle testing, shelf life studies, e.t.c.. Further uses of BOD incubator extend to microorganisms cultivation for biological studies, refrigerated storage for botany, sewage and water pollution.
We envision to be one of the most efficient manufacturers of Temperature and Humidity base equipment by adopting advanced techniques for enhancing quality and constantly updating our better to best.
We manufacture all our equipment as per GMP / ICH guidelines and also confirm MCA & US FDA requirements, so you don't need to worry about the quality of the product as we fill all the boxes of quality assurance.
You can visit us at https://www.kesarcontrol.com/ to contact us, to place an order and to get more information about what we do and how we do.

For more details about our product visit:   https://www.kesarcontrol.com/stability-chamber.php